requirements.txt
aiohttp==3.8.5
aiosignal==1.2.0
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.2
attrs==23.1.0
BeautifulReport==0.1.3
billiard==4.2.0
Brotli==1.0.9
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
confluent-kafka==2.4.0
cryptography==41.0.3
ddt==1.7.0
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.14.0
dnspython==2.4.2
eventlet==0.33.3
exceptiongroup==1.2.0
frozenlist==1.4.0
gevent==23.9.1
greenlet==3.0.1
gunicorn==20.1.0
HTMLReport==2.4.0
idna==2.10
iniconfig==2.0.0
Jinja2==3.0.3
jsonpath==0.82.2
kombu==5.3.4
MarkupSafe==2.1.3
multidict==6.0.2
mysqlclient==2.0.3
packaging==23.2
pip==23.3
pluggy==1.3.0
prettytable==3.10.0
prompt-toolkit==3.0.41
pycparser==2.21
pydantic==1.10.12
pyecharts==2.0.5
PyMySQL==1.1.0
pyOpenSSL==23.2.0
PySocks==1.7.1
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==5.3.1
redis==5.0.1
requests==2.24.0
setuptools==68.0.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
supervisor==4.2.2
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
unittestreport==1.5.1
urllib3==1.25.11
uWSGI==2.0.21
vine==5.1.0
wcwidth==0.2.12
wheel==0.41.2
wrapt==1.16.0
yarl==1.8.1
zope.event==5.0
zope.interface==5.4.0
